+++ NEWS.md
@@ -21,8 +21,12 @@
   predictions should be on the response scale, link scale, etc.). It can also
   be used to predict auxiliary (distributional) parameters.
 
-- The `"marginaleffects"` backend for `estimate_means()` is now fully implemented
-  and no longer work-in-progress.
+- `estimate_means()` and `estimate_contrasts()` get a ``marginalize` argument,
+  to specify how to marginalize over non-focal terms. This results in slightly
+  different predicted values, each approach answering a different question.
+
+- The `"marginaleffects"` backend is now fully implemented and no longer
+  work-in-progress.
 
 - `estimate_contrasts()` gains a `backend` argument. This defaults to `"emmeans"`,
   but can be set to `"marginaleffects"` to use features of that package to estimate